requirements.txt
Pillow
django-environ
selenium
chromedriver-py
